+++ packages/ruleset-bundler/src/plugins/__tests__/skypack.spec.ts
@@ -2,8 +2,6 @@ import * as fs from 'fs';
 import { serveAssets } from '@stoplight/spectral-test-utils';
 import { fetch } from '@stoplight/spectral-runtime';
 
-jest.mock?.('fs');
-
 import { BundleOptions, bundleRuleset } from '../../index';
 import type { IO } from '../../types';
 import { virtualFs } from '../virtualFs';
@@ -11,12 +9,21 @@ import { skypack } from '../skypack';
 
 describe('Skypack Plugin', () => {
   let io: IO;
+  let warnSpy: jest.SpyInstance;
 
   beforeEach(() => {
     io = {
       fs,
       fetch,
     };
+
+    warnSpy = jest.spyOn(console, 'warn').mockImplementation(() => {
+      /* no-op */
+    });
+  });
+
+  afterEach(() => {
+    warnSpy.mockRestore();
   });
 
   describe.each<BundleOptions['target']>(['browser'])('given %s target', target => {